Move AccountDropdown into Header folder
