+++ ui/src/Header/AccountDropdown/AccountDropdown.scss
@@ -15,8 +15,8 @@
  * limitations under the License.
  */
 
-@import '../Styles/Colors.scss';
-@import '../Styles/Breakpoints.scss';
+@import '../../Styles/Colors';
+@import '../../Styles/Breakpoints';
 
 .accountDropdownToggle {
   background: white;
