+++ ui/src/Header/AccountDropdown/ShareAccessButton/ShareAccessButton.tsx
