+++ ui/src/Header/AccountDropdown/ShareAccessForm/GrantAccessConfirmationForm/GrantEditAccessConfirmationForm.scss
@@ -16,7 +16,7 @@
  */
 
 
-@import "Styles/Colors";
+@import "../../../../Styles/Colors";
 
 .grant-edit-access-confirmation-form {
   margin: 0 auto;
