+++ ui/src/Header/AccountDropdown/ShareAccessForm/InviteEditorsFormSection/InviteEditorsFormSection.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import "Styles/Colors";
+@import "../../../../Styles/Colors";
 
 .inviteEditorsForm {
   width: 415px;
