+++ ui/src/Header/AccountDropdown/ShareAccessForm/InviteEditorsFormSection/UserAccessList/UserAccessList.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import '../../../../Styles/Colors';
+@import '../../../../../Styles/Colors';
 
 .userAccess-dropdown {
   font-size: 0.875rem;
