+++ ui/src/Header/AccountDropdown/ShareAccessForm/MultiModalCardBanner/MultiModalCardBanner.scss
@@ -15,9 +15,9 @@
  * limitations under the License.
  */
 
-@import '../../../Styles/Variables';
-@import '../../../Styles/Colors';
-@import '../../../Styles/Breakpoints';
+@import '../../../../Styles/Variables';
+@import '../../../../Styles/Colors';
+@import '../../../../Styles/Breakpoints';
 
 .modalCardBanner.multiModal{
   margin: 0 auto;
