+++ ui/src/Header/AccountDropdown/ShareAccessForm/ShareAccessForm.scss
