+++ ui/src/Header/AccountDropdown/ShareAccessForm/ShareAccessForm.test.tsx
@@ -16,14 +16,14 @@
  */
 
 import ShareAccessForm from './ShareAccessForm';
-import {renderWithRecoil} from '../../Utils/TestUtils';
+import {renderWithRecoil} from '../../../Utils/TestUtils';
 import React from 'react';
-import TestData from '../../Utils/TestData';
-import SpaceClient from '../../Services/Api/SpaceClient';
+import TestData from '../../../Utils/TestData';
+import SpaceClient from '../../../Services/Api/SpaceClient';
 import {screen, waitFor} from '@testing-library/react';
-import {RecoilObserver} from '../../Utils/RecoilObserver';
-import {ModalContents, ModalContentsState} from '../../State/ModalContentsState';
-import {CurrentSpaceState} from '../../State/CurrentSpaceState';
+import {RecoilObserver} from '../../../Utils/RecoilObserver';
+import {ModalContents, ModalContentsState} from '../../../State/ModalContentsState';
+import {CurrentSpaceState} from '../../../State/CurrentSpaceState';
 
 jest.mock('Services/Api/SpaceClient');
