+++ ui/src/Header/AccountDropdown/ShareAccessForm/ViewOnlyAccessFormSection/ViewOnlyAccessFormSection.scss
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-@import "Styles/Colors";
+@import "../../../../Styles/Colors";
 
 .viewOnlyAccessForm {
   width: 415px;
