+++ ui/src/Header/AccountDropdown/SignOutButton/SignOutButton.tsx
@@ -15,11 +15,11 @@
  * limitations under the License.
  */
 
-import {removeToken} from '../Services/TokenService';
+import {removeToken} from '../../../Services/TokenService';
 import React from 'react';
 import {useNavigate} from 'react-router-dom';
 import {useSetRecoilState} from 'recoil';
-import {CurrentUserState} from '../State/CurrentUserState';
+import {CurrentUserState} from '../../../State/CurrentUserState';
 
 interface Props {
     focusOnRender?: boolean;
