+++ ui/src/Header/Header.tsx
@@ -17,7 +17,7 @@
 
 import React from 'react';
 import PeopleMoverLogo from '../Common/PeopleMoverLogo/PeopleMoverLogo';
-import AccountDropdown from '../AccountDropdown/AccountDropdown';
+import AccountDropdown from './AccountDropdown/AccountDropdown';
 import {Link, useLocation} from 'react-router-dom';
 import {useRecoilValue} from 'recoil';
 import {CurrentSpaceState, UUIDForCurrentSpaceSelector} from '../State/CurrentSpaceState';
